[FEATURE] Improve type_overview #306
- added new sorting modes

- moved code to anndata_ext

- refactored type_overview code
