+++ ehrapy/__init__.py
@@ -14,7 +14,7 @@ traceback.install(width=200, word_wrap=True)
 
 from ehrapy._settings import EhrapyConfig, ehrapy_settings
 from ehrapy._util import print_versions
-from ehrapy.anndata_ext import anndata_to_df, df_to_anndata
+from ehrapy.anndata_ext import df_to_anndata, anndata_to_df, type_overview
 
 settings: EhrapyConfig = ehrapy_settings
